+++ frontend/src/features/dashboard/DashboardCalendar/Calendar.tsx
@@ -7,7 +7,7 @@ import palette from "styles/colorPalette";
 import { useDispatch, useSelector } from "react-redux";
 import { dashboardActions } from "slices/dashboardSlice";
 import { rootState } from "app/store";
-import api from "api/Api";
+import api from "api/BaseUrl";
 import Axios from "api/JsonAxios";
 
 const CalendarStyled = styled(Calendar)(() => ({
