+++ frontend/src/features/dashboard/DashboardList.tsx
@@ -4,7 +4,7 @@ import Text from "components/Text";
 import { useDispatch, useSelector } from "react-redux";
 import { dashboardActions } from "slices/dashboardSlice";
 import { rootState } from "app/store";
-import api from "api/Api";
+import api from "api/BaseUrl";
 import Axios from "api/JsonAxios";
 import { Link } from "@mui/material";
