+++ frontend/src/features/post/PageViewerContents.tsx
@@ -11,7 +11,7 @@ import { setPostContent, setPostFileList, setPostFiles } from "slices/postSlice"
 import { useSelector } from "react-redux";
 import { rootState } from "app/store";
 import Axios from "api/MultipartAxios";
-import api from "api/Api";
+import api from "api/BaseUrl";
 
 interface PageViewerContentsProps {
   content: string;
