+++ frontend/src/features/post/PostList.tsx
@@ -15,7 +15,7 @@ import {
 import { useNavigate } from "react-router-dom";
 import Axios from "api/JsonAxios";
 import { useSelector } from "react-redux";
-import api from "api/Api";
+import api from "api/BaseUrl";
 import { rootState } from "app/store";
 import { CircularProgress } from "@mui/material";
